rewrite rewrite nl802154_get_frequency and nl802154_channel2freq functions

+++ include/iwpaninfo.h
@@ -37,7 +37,7 @@ struct iwpaninfo_txpwrlist_entry {
 
 struct iwpaninfo_freqlist_entry {
 	uint8_t channel;
-	uint32_t mhz;
+	float mhz;
 };
 
 struct iwpaninfo_ops {
